Moved some basic arrays to use std containers.
This PR is the start of moving from basic C-style arrays to
using C++ STL containers. This move will help reduce the
number of NREs and out of bounds issues. It will also
open the code up to using some more advanced library
functions that are in the STL (eg. std::find_if from <algorithm>).