simplify lesser_demon_tick logic for MOB_VNUM_AAMON
This simple change makes the code more readable, making it clear that
there is no mistake or bad side effects due to the positioning of the
break. It also avoids unnecessary checks, optimizing the code.

The entire switch could be refactored, and perhaps no longer use ifs
within but instead have inner switches as necessary, but, for now,
with this change the case for MOB_VNUM_AAMON keeps the same stile we
see in the rest of the cases within this switch.

Closes: #301